==> 02_projeto.ipynb <==
import pandas as pd

vendas = pd.read_csv('01_projeto.csv', sep=';')
vendas

# Conversão da coluna de data para o formato datetime
vendas['data_venda'] = pd.to_datetime(vendas['data_venda'])    

# Criação da coluna com o dia do mês
vendas['dia_mes'] = vendas['data_venda'].dt.day

# Criação da coluna com o nome do dia da semana (em português)
vendas['dia'] = vendas['data_venda'].dt.day_name(locale='pt_BR')

# Criação da coluna com o número do dia da semana (0 = segunda, 6 = domingo)
vendas['dia_semana'] = vendas['data_venda'].dt.weekday

# Criação da coluna com o trimestre da venda
vendas['trimestre'] = vendas['data_venda'].dt.quarter

# Criação da coluna com o número do mês
vendas['mes_num'] = vendas['data_venda'].dt.month

# -------------------------------------------------
# Análises Semanais
# -------------------------------------------------

# Quantidade total de vendas por dia da semana
qnt_dia = vendas.groupby(['dia', 'dia_semana'])[['quantidade']].sum()
tot_dia = qnt_dia.sort_values(by='dia_semana' , ascending=True)
tot_dia

# Dia da semana que teve maior quantidade vendida
qnt_max = vendas.groupby(['dia', 'dia_semana'])[['quantidade']].sum()
qnt_max[qnt_max['quantidade'] == qnt_max['quantidade'].max()]

# Faturamento total por dia da semana
fat_dia = vendas.groupby(['dia', 'dia_semana'])[['ValorDaVenda']].sum()
fat_dia = fat_dia.sort_values(by='dia_semana' , ascending=True)
fat_dia

# Dia da semana que teve maior faturamento
fat_max = vendas.groupby(['dia', 'dia_semana'])[['ValorDaVenda']].sum()
fat_max[fat_max['ValorDaVenda'] == fat_max['ValorDaVenda'].max()]

# Lucro total por dia da semana
lucro_dia = vendas.groupby(['dia', 'dia_semana'])[['ValorDoLucro']].sum()
lucro_dia_dia = lucro_dia.sort_values(by='dia_semana' , ascending=True)
lucro_dia

# Dia da semana que teve maior lucro
lucro_max = vendas.groupby(['dia', 'dia_semana'])[['ValorDoLucro']].sum()
lucro_max[lucro_max['ValorDoLucro'] == lucro_max['ValorDoLucro'].max()]

# -------------------------------------------------
# Análises Mensais
# -------------------------------------------------

# Quantidade vendida por mês
qnt_mes = vendas.groupby(['mes', 'mes_num'])[['quantidade']].sum()
qnt_mes = qnt_mes.sort_values(by='mes_num', ascending=True)
qnt_mes

# Lucro total por mês (ordenado)
lucro_mes = vendas.groupby(['mes', 'mes_num'])[['ValorDoLucro']].sum()
lucro_mes = lucro_mes.sort_values(by='mes_num', ascending=True)
lucro_mes

# Mês com maior lucro
lucro_mes = vendas.groupby(['mes', 'mes_num'])[['ValorDoLucro']].sum()
lucro_mes[lucro_mes['ValorDoLucro'] == lucro_mes['ValorDoLucro'].max()]

# Mês com menor lucro
lucro_mes = vendas.groupby(['mes', 'mes_num'])[['ValorDoLucro']].sum()
lucro_mes[lucro_mes['ValorDoLucro'] == lucro_mes['ValorDoLucro'].min()]

# Faturamento total por mês
fat_mes = vendas.groupby(['mes', 'mes_num'])[['ValorDaVenda']].sum()
fat_mes = fat_mes.sort_values(by='mes_num', ascending=True)
fat_mes

# Mês de maior faturamento
fat_mes = vendas.groupby(['mes', 'mes_num'])[['ValorDaVenda']].sum()
fat_mes[fat_mes['ValorDaVenda'] == fat_mes['ValorDaVenda'].max()]

# Mês de menor faturamento
fat_mes = vendas.groupby(['mes', 'mes_num'])[['ValorDaVenda']].sum()
fat_mes[fat_mes['ValorDaVenda'] == fat_mes['ValorDaVenda'].min()]

# -------------------------------------------------
# Análises Trimestrais
# -------------------------------------------------

# Lucro total por trimestre
lucro_tri = vendas.groupby('trimestre')[['ValorDoLucro']].sum()
lucro_tri = lucro_tri.sort_values(by='trimestre', ascending=True)
lucro_tri

# Faturamento total por trimestre
fat_tri = vendas.groupby('trimestre')[['ValorDaVenda']].sum()
fat_tri = fat_tri.sort_values(by='trimestre', ascending=True)
fat_tri

# Quantidade total por trimestre
qnt_tri = vendas.groupby('trimestre')[['quantidade']].sum()
qnt_tri = qnt_tri.sort_values(by='trimestre', ascending=True)
qnt_tri

# Exporta o arquivo com as novas colunas e análises
vendas.to_csv('02_projeto.csv', index=False, sep=';', decimal=',', encoding='utf-8-sig')
