==> 03_projeto.ipynb <==

import pandas as pd 

vendas = pd.read_csv('02_projeto.csv', sep=';', decimal=',')

# -------------------------------------------------------------
# TICKET MÉDIO GERAL
# -------------------------------------------------------------

# Garante que a coluna 'quantidade' está no formato numérico (float)
# Isso evita erros ao fazer cálculos como soma e divisão.
vendas['quantidade'] = vendas['quantidade'].astype(float, copy=False)

# Calcula o TICKET MÉDIO GERAL:
# Fórmula = faturamento total / quantidade total vendida
ticket_medio = vendas['ValorDaVenda'].sum() / vendas['quantidade'].sum()

# -------------------------------------------------------------
# TICKET MÉDIO MENSAL
# -------------------------------------------------------------

# Agrupa por mês e calcula a soma do faturamento e da quantidade
ticket_mes = vendas.groupby('mes')[['ValorDaVenda', 'quantidade']].sum()

# Calcula o ticket médio por mês
ticket_mes['ticket_medio_mes'] = ticket_mes['ValorDaVenda'] / ticket_mes['quantidade']

# Organiza da maior para a menor margem
ticket_medio_mes = ticket_mes.sort_values(by='ticket_medio_mes', ascending=False)

# -------------------------------------------------------------
# TICKET MÉDIO POR DIA
# -------------------------------------------------------------

ticket_dia = vendas.groupby('dia')[['ValorDaVenda', 'quantidade']].sum()
ticket_dia['ticket_medio_dia'] = ticket_dia['ValorDaVenda'] / ticket_dia['quantidade']
ticket_medio_dia = ticket_dia.sort_values(by='ticket_medio_dia', ascending=False)

# -------------------------------------------------------------
# TICKET MÉDIO POR PRODUTO
# -------------------------------------------------------------

ticket_produto = vendas.groupby('produto')[['ValorDaVenda', 'quantidade']].sum()
ticket_produto['ticket_medio_prod'] = ticket_produto['ValorDaVenda'] / ticket_produto['quantidade']
ticket_medio_prod = ticket_produto.sort_values(by='ticket_medio_prod', ascending=False)

# -------------------------------------------------------------
# TICKET MÉDIO POR CATEGORIA
# -------------------------------------------------------------

ticket_categoria = vendas.groupby('categoria')[['ValorDaVenda', 'quantidade']].sum()
ticket_categoria['ticket_medio_categ'] = ticket_categoria['ValorDaVenda'] / ticket_categoria['quantidade']
ticket_medio_categ = ticket_categoria.sort_values(by='ticket_medio_categ', ascending=False)

# -------------------------------------------------------------
# MARGEM DE LUCRO POR MÊS
# -------------------------------------------------------------
# Fórmula da margem:
# Margem (%) = (Lucro / Faturamento) * 100

lucro_mes = vendas.groupby('mes')[['ValorDoLucro', 'ValorDaVenda']].sum()
lucro_mes['margem_lucro_mes'] = (lucro_mes['ValorDoLucro'] / lucro_mes['ValorDaVenda']) * 100
margem_lucro_mes = lucro_mes.sort_values(by='mes', ascending=False)

# -------------------------------------------------------------
# MARGEM DE LUCRO POR DIA
# -------------------------------------------------------------

lucro_dia = vendas.groupby('dia')[['ValorDoLucro', 'ValorDaVenda']].sum()
lucro_dia['margem_lucro_dia'] = (lucro_dia['ValorDoLucro'] / lucro_dia['ValorDaVenda']) * 100
margem_lucro_dia = lucro_dia.sort_values(by='dia', ascending=False)

# -------------------------------------------------------------
# MARGEM DE LUCRO POR PRODUTO
# -------------------------------------------------------------

lucro_produto = vendas.groupby('produto')[['ValorDoLucro', 'ValorDaVenda']].sum()
lucro_produto['margem_lucro_produto'] = (lucro_produto['ValorDoLucro'] / lucro_produto['ValorDaVenda']) * 100
margem_lucro_produto = lucro_produto.sort_values('produto', ascending=False)

# -------------------------------------------------------------
# MARGEM DE LUCRO POR CATEGORIA
# -------------------------------------------------------------

lucro_categoria = vendas.groupby('categoria')[['ValorDoLucro', 'ValorDaVenda']].sum()
lucro_categoria['margem_lucro_categoria'] = (lucro_categoria['ValorDoLucro'] / lucro_categoria['ValorDaVenda']) * 100
margem_lucro_categoria = lucro_categoria.sort_values('categoria', ascending=False)
