==> 04_projeto.ipynb <==
import pandas as pd 

vendas = pd.read_csv('02_projeto.csv', sep=';', decimal=',')
vendas.head()

total_lucro = sum(vendas['ValorDoLucro'])
lucro_produto = vendas.groupby('produto')[['ValorDoLucro']].sum().sort_values(by = 'ValorDoLucro', ascending=False)
porc_lucro_produto = (lucro_produto / total_lucro) * 100
porc_lucro_produto


lucro_categoria = vendas.groupby('categoria')[['ValorDoLucro']].sum().sort_values(by = 'ValorDoLucro', ascending=False)
porc_lucro_categoria = (lucro_categoria / total_lucro) * 100
porc_lucro_categoria
